+++ capa/features/extractors/viv/insn.py
@@ -271,6 +271,10 @@ def extract_insn_bytes_features(fh: FunctionHandle, bb, ih: InsnHandle) -> Itera
             if capa.features.extractors.helpers.all_zeros(buf):
                 continue
 
+            if f.vw.isProbablyString(v):
+                # don't extract byte features for obvious strings
+                continue
+
             yield Bytes(buf), ih.address